+++ src/core/registry.js
@@ -17,19 +17,16 @@
  * - set pattern.jquery_plugin if you want it
  */
 import $ from "jquery";
-import _ from "underscore";
+import dom from "./dom";
 import logging from "./logging";
 import utils from "./utils";
 
-// below here modules that are only loaded
-import "./jquery-ext";
-
-var log = logging.getLogger("registry"),
-    disable_re = /patterns-disable=([^&]+)/g,
-    dont_catch_re = /patterns-dont-catch/g,
-    dont_catch = false,
-    disabled = {},
-    match;
+const log = logging.getLogger("registry");
+const disable_re = /patterns-disable=([^&]+)/g;
+const dont_catch_re = /patterns-dont-catch/g;
+const disabled = {};
+let dont_catch = false;
+let match;
 
 while ((match = disable_re.exec(window.location.search)) !== null) {
     disabled[match[1]] = true;
@@ -41,14 +38,14 @@ while ((match = dont_catch_re.exec(window.location.search)) !== null) {
     log.info("I will not catch init exceptions");
 }
 
-var registry = {
+const registry = {
     patterns: {},
     // as long as the registry is not initialized, pattern
     // registration just registers a pattern. Once init is called,
     // the DOM is scanned. After that registering a new pattern
     // results in rescanning the DOM only for this pattern.
     initialized: false,
-    init: function registry_init() {
+    init() {
         $(document).ready(function () {
             log.info(
                 "loaded: " + Object.keys(registry.patterns).sort().join(", ")
@@ -59,13 +56,13 @@ var registry = {
         });
     },
 
-    clear: function clearRegistry() {
+    clear() {
         // Removes all patterns from the registry. Currently only being
         // used in tests.
         this.patterns = {};
     },
 
-    transformPattern: function (name, content) {
+    transformPattern(name, content) {
         /* Call the transform method on the pattern with the given name, if
          * it exists.
          */
@@ -88,14 +85,14 @@ var registry = {
         }
     },
 
-    initPattern: function (name, el, trigger) {
+    initPattern(name, el, trigger) {
         /* Initialize the pattern with the provided name and in the context
          * of the passed in DOM element.
          */
-        var $el = $(el);
-        var pattern = registry.patterns[name];
+        const $el = $(el);
+        const pattern = registry.patterns[name];
         if (pattern.init) {
-            var plog = logging.getLogger("pat." + name);
+            const plog = logging.getLogger("pat." + name);
             if ($el.is(pattern.trigger)) {
                 plog.debug("Initialising:", $el);
                 try {
@@ -111,54 +108,61 @@ var registry = {
         }
     },
 
-    orderPatterns: function (patterns) {
+    orderPatterns(patterns) {
         // XXX: Bit of a hack. We need the validation pattern to be
         // parsed and initiated before the inject pattern. So we make
         // sure here, that it appears first. Not sure what would be
         // the best solution. Perhaps some kind of way to register
         // patterns "before" or "after" other patterns.
-        if (
-            _.contains(patterns, "validation") &&
-            _.contains(patterns, "inject")
-        ) {
+        if (patterns.includes("validation") && patterns.includes("inject")) {
             patterns.splice(patterns.indexOf("validation"), 1);
             patterns.unshift("validation");
         }
         return patterns;
     },
 
-    scan: function registryScan(content, patterns, trigger) {
-        var selectors = [],
-            $match;
+    scan(content, patterns, trigger) {
+        if (typeof content === "string") {
+            content = document.querySelector(content);
+        } else if (content.jquery) {
+            content = content[0];
+        }
+
+        const selectors = [];
         patterns = this.orderPatterns(
             patterns || Object.keys(registry.patterns)
         );
-        patterns.forEach(_.partial(this.transformPattern, _, content));
-        patterns = _.each(patterns, function (name) {
-            var pattern = registry.patterns[name];
+        for (const name of patterns) {
+            this.transformPattern(name, content);
+            const pattern = registry.patterns[name];
             if (pattern.trigger) {
                 selectors.unshift(pattern.trigger);
             }
-        });
-        $match = $(content).findInclusive(selectors.join(",")); // Find all DOM elements belonging to a pattern
-        $match = $match.filter(function () {
+        }
+
+        let matches = dom.querySelectorAllAndMe(
+            content,
+            selectors.map((it) => it.trim().replace(/,$/, "")).join(",")
+        );
+        matches = matches.filter((el) => {
             // Filter out code examples wrapped in <pre> elements.
-            return $(this).parents("pre").length === 0;
+            // Also filter special class ``.cant-touch-this``
+            return (
+                dom.find_parents(el, "pre").length === 0 &&
+                !el.matches(".cant-touch-this")
+            );
         });
-        $match = $match.filter(":not(.cant-touch-this)");
 
         // walk list backwards and initialize patterns inside-out.
-        $match.toArray().reduceRight(
-            function registryInitPattern(acc, el) {
-                patterns.forEach(_.partial(this.initPattern, _, el, trigger));
-            }.bind(this),
-            null
-        );
-        $("body").addClass("patterns-loaded");
+        for (const el of matches.reverse()) {
+            for (const name of patterns) {
+                this.initPattern(name, el, trigger);
+            }
+        }
+        document.body.classList.add("patterns-loaded");
     },
 
-    register: function registry_register(pattern, name) {
-        var plugin_name;
+    register(pattern, name) {
         name = name || pattern.name;
         if (!name) {
             log.error("Pattern lacks a name:", pattern);
@@ -174,12 +178,12 @@ var registry = {
 
         // register pattern as jquery plugin
         if (pattern.jquery_plugin) {
-            plugin_name = ("pat-" + name).replace(/-([a-zA-Z])/g, function (
-                match,
-                p1
-            ) {
-                return p1.toUpperCase();
-            });
+            const plugin_name = ("pat-" + name).replace(
+                /-([a-zA-Z])/g,
+                function (match, p1) {
+                    return p1.toUpperCase();
+                }
+            );
             $.fn[plugin_name] = utils.jqueryPlugin(pattern);
             // BBB 2012-12-10 and also for Mockup patterns.
             $.fn[plugin_name.replace(/^pat/, "pattern")] = $.fn[plugin_name];